refactor(hub): descomponer god object app.go en archivos por dominio
Split de app.go (1166 líneas, ~35 métodos) en 7 archivos organizados
por dominio:

- app.go: struct App, tipos, NewApp, startup, shutdown (~143 líneas)
- app_discovery.go: mDNS discovery (~90 líneas)
- app_connection.go: conexión y pairing WebSocket (~262 líneas)
- app_upload.go: upload de juegos y artwork local (~330 líneas)
- app_games.go: juegos remotos (~89 líneas)
- app_steamgriddb.go: SteamGridDB + artwork helpers (~147 líneas)
- app_settings.go: settings, cache, identidad, version (~135 líneas)

Bonus: extraer helper steamgriddbClient() para eliminar duplicación
5x del patrón apiKey + NewClient en los métodos de SteamGridDB.

Sin cambios de comportamiento — solo reorganización de código.

Closes #67
